requirements.txt
streamlit==1.45.1
numpy==2.2.6
faiss-cpu
sentence-transformers==2.6.1
openai
PyMuPDF==1.23.26
google-generativeai
transformers
torch
huggingface-hub==0.23.1
protobuf==4.25.3
